requirements.txt
# MoonGuard - מערכת אימות רכבים
# Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# File Watching
watchdog==3.0.0

# AI Integration
openai==1.12.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Database
aiosqlite==0.19.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.3

